fix: consolidate all QR scanning code into one class, prevent the process command from getting called unnecessarily, general refactoring of QR scanning code
